+++ gmailconnector/__init__.py
@@ -12,14 +12,14 @@ from .send_email import SendEmail  # noqa: F401
 from .send_sms import SendSMS  # noqa: F401
 from .validator.validate_email import validate_email  # noqa: F401
 
-version = "0.7.5"
+version = "0.7.6"
 
 
 def load_env(filename: Union[str, os.PathLike] = ".env", scan: bool = False) -> NoReturn:
     """Load .env files."""
     if scan:
         for file in os.listdir():
-            if file.endswith(".env"):
+            if os.path.isfile(file) and file.endswith(".env"):
                 dotenv.load_dotenv(dotenv_path=file, verbose=False)
     else:
         if os.path.isfile(filename):